Move Basic detail information view to its own component
